requirements.txt
dotenv==0.9.9
python-telegram-bot==22.0
asyncio==3.4.3
SQLAlchemy==2.0.40
boto3==1.37.33